# docs/openapi.json
{
  "components": {
    "schemas": {
      "AccrualMethod": {
        "enum": [
          "TIME",
          "HOURS_WORKED"
        ],
        "title": "AccrualMethod",
        "type": "string"
      },
...
      "AccrualRuleOut": {
        "properties": {
          "accrues_at": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccruesAt"
              },
              {
                "type": "null"
              }
            ]
          },
          "amount": {
            "title": "Amount",
            "type": "string"
          },
          "frequency": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Schedule"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "method": {
            "$ref": "#/components/schemas/AccrualMethod"
          },
          "min_tenure_months": {
            "title": "Min Tenure Months",
            "type": "integer"
          },
          "per_minutes_worked": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Per Minutes Worked"
          },
          "unit": {
            "$ref": "#/components/schemas/RateUnit"
          }
        },
        "required": [
          "id",
          "method",
          "amount",
          "unit",
          "frequency",
          "accrues_at",
          "per_minutes_worked",
          "min_tenure_months"
        ],
        "title": "AccrualRuleOut",
        "type": "object"
      },
      "AccruesAt": {
        "enum": [
          "START_OF_PERIOD",
          "END_OF_PERIOD"
        ],
        "title": "AccruesAt",
        "type": "string"
      },
...
      "EmployeeGroupMembershipIn": {
        "properties": {
          "effective_from": {
            "format": "date",
            "title": "Effective From",
            "type": "string"
          },
          "group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          }
        },
        "required": [
          "effective_from"
        ],
        "title": "EmployeeGroupMembershipIn",
        "type": "object"
      },
      "EmployeeGroupOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/GroupMemberOut"
            },
            "title": "Members",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "members"
        ],
        "title": "EmployeeGroupOut",
        "type": "object"
      },
...
      "GroupCreateIn": {
        "properties": {
          "employee_ids": {
            "items": {
              "type": "string"
            },
            "title": "Employee Ids",
            "type": "array"
          },
          "name": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "GroupCreateIn",
        "type": "object"
      },
      "GroupMemberOut": {
        "properties": {
          "employee_id": {
            "title": "Employee Id",
            "type": "string"
          },
          "employee_name": {
            "title": "Employee Name",
            "type": "string"
          },
          "employment_type": {
            "title": "Employment Type",
            "type": "string"
          }
        },
        "required": [
          "employee_id",
          "employee_name",
          "employment_type"
        ],
        "title": "GroupMemberOut",
        "type": "object"
      },
      "GroupMembersIn": {
        "properties": {
          "effective_from": {
            "format": "date",
            "title": "Effective From",
            "type": "string"
          },
          "employee_ids": {
            "items": {
              "type": "string"
            },
            "title": "Employee Ids",
            "type": "array"
          }
        },
        "required": [
          "effective_from"
        ],
        "title": "GroupMembersIn",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NewHireProration": {
        "enum": [
          "PRORATE",
          "FULL",
          "NONE"
        ],
        "title": "NewHireProration",
        "type": "string"
      },
...
      "PolicyAudienceIn": {
        "properties": {
          "all_employees": {
            "default": false,
            "title": "All Employees",
            "type": "boolean"
          },
          "effective_from": {
            "format": "date",
            "title": "Effective From",
            "type": "string"
          },
          "group_ids": {
            "items": {
              "type": "string"
            },
            "title": "Group Ids",
            "type": "array"
          }
        },
        "required": [
          "effective_from"
        ],
        "title": "PolicyAudienceIn",
        "type": "object"
      },
...
      "PolicyKind": {
        "enum": [
          "UNLIMITED",
          "ACCRUAL"
        ],
        "title": "PolicyKind",
        "type": "string"
      },
      "PolicyOut": {
        "properties": {
          "all_employees": {
            "title": "All Employees",
            "type": "boolean"
          },
          "category_id": {
            "title": "Category Id",
            "type": "string"
          },
          "category_name": {
            "title": "Category Name",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "current_version": {
            "$ref": "#/components/schemas/PolicyVersionOut"
          },
          "group_ids": {
            "items": {
              "type": "string"
            },
            "title": "Group Ids",
            "type": "array"
          },
          "group_names": {
            "items": {
              "type": "string"
            },
            "title": "Group Names",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "version_count": {
            "title": "Version Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "category_id",
          "category_name",
          "created_by",
          "current_version",
          "version_count",
          "all_employees",
          "group_ids",
          "group_names"
        ],
        "title": "PolicyOut",
        "type": "object"
      },
...
      "PolicyVersionOut": {
        "properties": {
          "allow_negative": {
            "title": "Allow Negative",
            "type": "boolean"
          },
          "carryover_cap_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carryover Cap Minutes"
          },
          "change_reason": {
            "title": "Change Reason",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "effective_from": {
            "format": "date",
            "title": "Effective From",
            "type": "string"
          },
          "expires_at_period_end": {
            "title": "Expires At Period End",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/PolicyKind"
          },
          "max_balance_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Balance Minutes"
          },
          "negative_floor_minutes": {
            "title": "Negative Floor Minutes",
            "type": "integer"
          },
          "new_hire_proration": {
            "$ref": "#/components/schemas/NewHireProration"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/AccrualRuleOut"
            },
            "title": "Rules",
            "type": "array"
          },
          "tenure_transition": {
            "$ref": "#/components/schemas/TenureTransition"
          },
          "version_no": {
            "title": "Version No",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "version_no",
          "effective_from",
          "kind",
          "created_by",
          "change_reason",
          "new_hire_proration",
          "allow_negative",
          "negative_floor_minutes",
          "max_balance_minutes",
          "carryover_cap_minutes",
          "expires_at_period_end",
          "tenure_transition",
          "created_at",
          "rules"
        ],
        "title": "PolicyVersionOut",
        "type": "object"
      },
      "RateUnit": {
        "enum": [
          "DAY",
          "HOUR",
          "MINUTE"
        ],
        "title": "RateUnit",
        "type": "string"
      },
...
      "Schedule": {
        "enum": [
          "DAILY",
          "WEEKLY",
          "SEMIMONTHLY",
          "BIWEEKLY",
          "MONTHLY",
          "YEARLY"
        ],
        "title": "Schedule",
        "type": "string"
      },
...
      "TenureTransition": {
        "enum": [
          "NEXT_PERIOD"
        ],
        "title": "TenureTransition",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Employee time off tracking: policies, accrual, requests, balances, audit.",
    "title": "Warp Time Off",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/employees/{employee_id}/group": {
      "put": {
        "operationId": "set_employee_group_api_employees__employee_id__group_put",
        "parameters": [
          {
            "in": "path",
            "name": "employee_id",
            "required": true,
            "schema": {
              "title": "Employee Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "default": "adm_lindsey",
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeGroupMembershipIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Employee Group",
        "tags": [
          "groups"
        ]
      }
    },
...
    "/api/groups": {
      "get": {
        "operationId": "list_groups_api_groups_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "default": "adm_lindsey",
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EmployeeGroupOut"
                  },
                  "title": "Response List Groups Api Groups Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Groups",
        "tags": [
          "groups"
        ]
      },
      "post": {
        "operationId": "create_group_api_groups_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "default": "adm_lindsey",
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeGroupOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Group",
        "tags": [
          "groups"
        ]
      }
    },
    "/api/groups/{group_id}": {
      "delete": {
        "operationId": "delete_group_api_groups__group_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "default": "adm_lindsey",
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Group",
        "tags": [
          "groups"
        ]
      }
    },
    "/api/groups/{group_id}/members": {
      "put": {
        "operationId": "replace_group_members_api_groups__group_id__members_put",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "default": "adm_lindsey",
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupMembersIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeGroupOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace Group Members",
        "tags": [
          "groups"
        ]
      }
    },
...
    "/api/policies/{policy_id}/audience": {
      "put": {
        "operationId": "set_policy_audience_api_policies__policy_id__audience_put",
        "parameters": [
          {
            "in": "path",
            "name": "policy_id",
            "required": true,
            "schema": {
              "title": "Policy Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "default": "adm_lindsey",
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyAudienceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Policy Audience",
        "tags": [
          "groups"
        ]
      }
    },
...
  }
}